+++ CHANGES.rst
@@ -4,7 +4,8 @@ Changes
 Unreleased
 ----------
 
- - Fixed bug where serializer of djangorestframework crashed if used with ``OrderingFilter`` (gh-821)
+- Fixed bug where serializer of djangorestframework crashed if used with ``OrderingFilter`` (gh-821)
+- Fixed bug where latest() is not idempotent for identical ``history_date`` records (gh-861)
 
 3.0.0 (2021-04-16)
 ------------------