+++ CHANGES.rst
@@ -6,6 +6,7 @@ Unreleased
 
 - Fixed bug where serializer of djangorestframework crashed if used with ``OrderingFilter`` (gh-821)
 - Fixed `make format` so it works by using tox (gh-859)
+- Fixed bug where latest() is not idempotent for identical ``history_date`` records (gh-861)
 
 3.0.0 (2021-04-16)
 ------------------